Splitting DynamoDB template library into cli and core part to allow for smaller lambda bundle sizes
