+++ workspaces/templates-lib/packages/template-dynamodb-cli/.npmignore
@@ -1,0 +1,6 @@
+/*
+!/dist/
+!/bin/
+!/package.json
+!/README.md
+/dist/tsconfig.tsbuildinfo
