+++ workspaces/templates-lib/packages/template-dynamodb-cli/README.md
@@ -1,0 +1,5 @@
+# Goldstack DynamoDB Template Tools
+
+This library provides an API and CLI for building Node.js modules that wrap a DynamoDB Table. These tools allow deploying the infrastructure for the table as well as a way to connect with the table across multiple environments (such as dev, staging and production).
+
+For more information, see [DynamoDB Template](https://docs.goldstack.party/docs/templates/dynamodb).
