+++ workspaces/templates-lib/packages/template-dynamodb-cli/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@goldstack/template-dynamodb",
-  "version": "0.1.10",
+  "name": "@goldstack/template-dynamodb-cli",
+  "version": "0.4.0",
   "description": "Utilities for building modules for DynamoDB access.",
   "keywords": [
     "goldstack",
@@ -22,7 +22,7 @@
   "license": "MIT",
   "author": "Max Rohde",
   "sideEffects": false,
-  "main": "src/templateDynamoDB.ts",
+  "main": "src/templateDynamoDBCli.ts",
   "scripts": {
     "build": "yarn clean && yarn compile",
     "build:watch": "yarn clean && yarn compile-watch",
@@ -39,16 +39,15 @@
     "version:apply:force": "yarn version $@ && yarn version apply"
   },
   "dependencies": {
-    "@aws-sdk/util-dynamodb": "^3.87.0",
-    "@goldstack/infra": "0.3.36",
-    "@goldstack/infra-aws": "0.3.44",
-    "@goldstack/utils-esbuild": "0.4.0",
+    "@goldstack/template-dynamodb": "0.1.10",
+    "@goldstack/utils-cli": "0.2.13",
     "@goldstack/utils-package": "0.3.38",
-    "@goldstack/utils-package-config-embedded": "0.4.0",
+    "@goldstack/utils-package-config": "0.3.38",
+    "@goldstack/utils-sh": "0.4.31",
     "@goldstack/utils-terraform": "0.3.68",
-    "aws-sdk": "2.1055.0",
-    "testcontainers": "^8.10.0",
-    "umzug": "^3.1.1"
+    "@goldstack/utils-terraform-aws": "0.3.67",
+    "umzug": "^3.1.1",
+    "yargs": "^15.4.1"
   },
   "devDependencies": {
     "@goldstack/utils-docs-cli": "0.3.11",
@@ -64,6 +63,6 @@
     "typescript": "^4.6.4"
   },
   "publishConfig": {
-    "main": "dist/src/templateDynamoDB.js"
+    "main": "dist/src/templateDynamoDBCli.js"
   }
 }
