+++ workspaces/templates/packages/dynamodb/package.json
@@ -29,6 +29,7 @@
     "umzug": "^3.1.1"
   },
   "devDependencies": {
+    "@goldstack/template-dynamodb-cli": "0.4.0",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.33",
     "jest": "^28.1.0",
