+++ workspaces/templates/packages/dynamodb/src/template.ts
@@ -1,4 +1,4 @@
-import { run } from '@goldstack/template-dynamodb';
+import { run } from '@goldstack/template-dynamodb-cli';
 import { createMigrations } from './migrations';
 
 run({ args: process.argv, migrations: createMigrations() }).catch((e) => {
